requirements.txt
fastapi
uvicorn
pydantic
requests
google-generativeai
python-dotenv
beautifulsoup4
supabase
lxml
httpx
motor
anyio
